:sparkles: feat: trazendo o setor selecionado do editar
